# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant CRM SaaS API (Phase 1 foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/delete-account:
    delete:
      operationId: deleteUserAccount
      tags: [auth]
      summary: Permanently delete the authenticated user's account
      description: >
        Requires the authenticated user to type DELETE. The external Clerk
        user, every organization where they currently hold an owner role, all
        memberships, and their local identity are permanently removed. Records
        in organizations they do not own remain and user references are
        detached.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAccountDeletionConfirmation"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAccountDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "404":
          description: User not found
        "409":
          description: Account deletion is already in progress
        "502":
          description: Clerk deletion failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
...
  /orgs/{orgId}/accounts/{accountId}:
    get:
      operationId: getAccount
      tags: [crm]
      summary: Account detail with related contacts and opportunities
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountWithRelations"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [crm]
      summary: Update an account
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetail"
    delete:
      operationId: deleteAccount
      tags: [crm]
      summary: Soft delete an account (sets is_active false)
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AccountId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
...
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    UserAccountDeletionConfirmation:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [DELETE] }
    UserAccountDeletionResponse:
      type: object
      additionalProperties: false
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
...
    Opportunity:
      type: object
      required: [id, name, stage]
      properties:
        id: { type: string }
        accountId: { type: string }
        name: { type: string }
        stage: { type: string }
        probability: { type: ["number", "null"] }
        value: { type: ["string", "null"] }
        expectedCloseDate: { type: ["string", "null"] }
        forecastCategory: { type: ["string", "null"] }
        ownerUserId: { type: ["string", "null"], format: uuid }
        createdByUserId: { type: ["string", "null"], format: uuid }
        createdAt: { type: ["string", "null"] }
...
    FileAttachment:
      type: object
      required: [objectPath, name]
      properties:
        objectPath: { type: string }
        name: { type: string }
        size: { type: ["integer", "null"] }
        contentType: { type: ["string", "null"] }
        uploadedAt: { type: ["string", "null"] }
    AccountDetail:
      type: object
      required: [id, name, isActive]
      properties:
        id: { type: string }
        name: { type: string }
        industry: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        zip: { type: ["string", "null"] }
        annualRevenue: { type: ["string", "null"] }
        employeeCount: { type: ["integer", "null"] }
        healthScore: { type: ["string", "null"] }
        riskLevel: { type: ["string", "null"] }
        ltv: { type: ["string", "null"] }
        nextRenewalDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        ownerUserId: { type: ["string", "null"], format: uuid }
        createdByUserId: { type: ["string", "null"], format: uuid }
        metadata: { type: object, additionalProperties: true }
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileAttachment"
        createdAt: { type: ["string", "null"] }
        updatedAt: { type: ["string", "null"] }
    AccountWithRelations:
      allOf:
        - $ref: "#/components/schemas/AccountDetail"
        - type: object
          required: [contacts, opportunities]
          properties:
            contacts:
              type: array
              items:
                $ref: "#/components/schemas/Contact"
            opportunities:
              type: array
              items:
                $ref: "#/components/schemas/Opportunity"
...
    AccountUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        industry: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        zip: { type: ["string", "null"] }
        annualRevenue: { type: ["string", "null"] }
        employeeCount: { type: ["integer", "null"] }
        healthScore: { type: ["string", "null"] }
        riskLevel: { type: ["string", "null"] }
        nextRenewalDate: { type: ["string", "null"] }
        metadata: { type: object, additionalProperties: true }
        ownerUserId: { type: ["string", "null"], format: uuid }
    Contact:
      type: object
      required: [id, accountId, firstName, lastName, isActive]
      properties:
        id: { type: string }
        accountId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        department: { type: ["string", "null"] }
        seniority: { type: ["string", "null"] }
        reportsToContactId: { type: ["string", "null"] }
        isActive: { type: boolean }
        metadata: { type: object, additionalProperties: true }
        ownerUserId: { type: ["string", "null"], format: uuid }
        createdByUserId: { type: ["string", "null"], format: uuid }
        createdAt: { type: ["string", "null"] }
...